# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "passalong-server API",
    "version": "1.0.0-draft",
    "summary": "The contract between passalong-server and the passalong client.",
    "description": "Draft of the protocol spike (PLAN-00001). REST + JSON; item content travels as byte streams. The server never interprets an item's `meta` and never exposes a field taken from it. Every request is safe to repeat: each operation that is not a GET says in `x-passalong-replay` what the same request, sent again, is answered. `x-passalong-errors` lists the codes an operation can refuse with. See README.md and rewrite-session.md beside this file."
  },
  "servers": [
    {
      "url": "https://pass.example.net:8443"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
...
    "/v1/viewer": {
      "get": {
        "operationId": "getViewer",
        "summary": "The API key in use, and the server.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Viewer"
                }
              }
            }
          },
          "default": {
            "description": "A refusal. `code` is stable; the client decides from it alone whether to retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-passalong-errors": [
          "INVALID_REQUEST",
          "KEY_EXPIRED",
          "KEY_REVOKED",
          "RATE_LIMITED",
          "SERVICE_UNAVAILABLE",
          "UNAUTHENTICATED"
        ]
      }
    },
    "/v1/workspace": {
      "get": {
        "operationId": "getWorkspace",
        "summary": "The workspace the key belongs to, with its encryption state and header. `passalong encrypt --join` reads the header here.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "default": {
            "description": "A refusal. `code` is stable; the client decides from it alone whether to retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-passalong-errors": [
          "INVALID_REQUEST",
          "KEY_EXPIRED",
          "KEY_REVOKED",
          "RATE_LIMITED",
          "SERVICE_UNAVAILABLE",
          "UNAUTHENTICATED"
        ]
      }
    },
    "/v1/workspace/probe": {
      "post": {
        "operationId": "probeWrite",
        "summary": "Checks that the key may write, that no rewrite is open, and that staging is writable, without storing an item.",
        "responses": {
          "204": {
            "description": "Done"
          },
          "default": {
            "description": "A refusal. `code` is stable; the client decides from it alone whether to retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-passalong-errors": [
          "FORBIDDEN_ROLE",
          "INVALID_REQUEST",
          "KEY_EXPIRED",
          "KEY_REVOKED",
          "RATE_LIMITED",
          "REWRITE_IN_PROGRESS",
          "SERVICE_UNAVAILABLE",
          "UNAUTHENTICATED"
        ],
        "x-passalong-replay": "Safe to repeat: it changes nothing."
      }
    },
    "/v1/workspace/clean-staging": {
      "post": {
        "operationId": "cleanStaging",
        "summary": "Forgets expired uploads and outcomes, and removes what interrupted changes left behind. The janitor does the same unasked.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanStagingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanStagingResult"
                }
              }
            }
          },
          "default": {
            "description": "A refusal. `code` is stable; the client decides from it alone whether to retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-passalong-errors": [
          "FORBIDDEN_ROLE",
          "INVALID_REQUEST",
          "KEY_EXPIRED",
          "KEY_REVOKED",
          "RATE_LIMITED",
          "SERVICE_UNAVAILABLE",
          "UNAUTHENTICATED"
        ],
        "x-passalong-replay": "Safe to repeat: a second pass finds nothing."
      }
    },
...
    "/v1/rewrite": {
      "get": {
        "operationId": "getRewrite",
        "summary": "The open rewrite session.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewriteSession"
                }
              }
            }
          },
          "default": {
            "description": "A refusal. `code` is stable; the client decides from it alone whether to retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-passalong-errors": [
          "INVALID_REQUEST",
          "KEY_EXPIRED",
          "KEY_REVOKED",
          "NOT_FOUND",
          "RATE_LIMITED",
          "SERVICE_UNAVAILABLE",
          "UNAUTHENTICATED"
        ]
      },
      "post": {
        "operationId": "beginRewrite",
        "summary": "Takes the lease and opens an empty next generation. Ordinary writers get REWRITE_IN_PROGRESS from here; readers keep reading.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RewriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewriteSession"
                }
              }
            }
          },
          "default": {
            "description": "A refusal. `code` is stable; the client decides from it alone whether to retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-passalong-errors": [
          "FORBIDDEN_ROLE",
          "INVALID_REQUEST",
          "KEY_EXPIRED",
          "KEY_ID_MISMATCH",
          "KEY_REVOKED",
          "RATE_LIMITED",
          "REWRITE_ENDED",
          "REWRITE_IN_PROGRESS",
          "SERVICE_UNAVAILABLE",
          "UNAUTHENTICATED"
        ],
        "x-passalong-replay": "By the holder with the same kind and new key id: the live session, lease renewed. After that rewrite was committed: KEY_ID_MISMATCH. After it was aborted: REWRITE_ENDED, for good; begin again under a new key."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "pal_<key id>_<secret>",
        "description": "The key selects the workspace; no request names one."
      }
    },
    "schemas": {
      "ItemId": {
        "type": "string",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{12}$",
        "description": "`<creation time>-<content key>`, as in the passalong client's item schema. Proposed by the client; parsed strictly."
      },
      "KeyId": {
        "type": "string",
        "pattern": "^[0-9a-f]{1,64}$",
        "description": "The id of a workspace's data key. The server compares key ids and never sees a key."
      },
...
      "Bytes": {
        "type": "string",
        "pattern": "^[0-9]+$",
        "description": "A byte count, as a string: item sizes are 64-bit and JSON numbers are not."
      },
      "Opaque": {
        "type": "object",
        "additionalProperties": true,
        "description": "A document the client wrote and the server never looks into."
      },
...
      "RewriteRequest": {
        "type": "object",
        "required": [
          "kind",
          "newKeyId",
          "newHeader"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "migrate",
              "rotate"
            ]
          },
          "expectedKeyId": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/KeyId"
              },
              {
                "type": "null"
              }
            ],
            "description": "The current key id; null for the plaintext workspace a migration starts from."
          },
          "newKeyId": {
            "$ref": "#/components/schemas/KeyId"
          },
          "newHeader": {
            "$ref": "#/components/schemas/Opaque"
          }
        }
      },
...
      "RewriteSession": {
        "type": "object",
        "required": [
          "kind",
          "holder",
          "leaseExpiresAt",
          "newKeyId",
          "newHeader",
          "stagedIds",
          "sourceItems"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "migrate",
              "rotate"
            ]
          },
          "holder": {
            "type": "string",
            "description": "The public id of the API key that holds the lease."
          },
          "leaseExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "newKeyId": {
            "$ref": "#/components/schemas/KeyId"
          },
          "stagedIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemId"
            }
          },
          "sourceItems": {
            "type": "integer",
            "minimum": 0
          },
          "newHeader": {
            "$ref": "#/components/schemas/Opaque",
            "description": "The header `beginRewrite` brought: the new data key, wrapped under the new words, which the workspace gets on commit. Whoever resumes the session, the holder after a restart or another device after `takeOverRewrite`, unlocks it with the new words. `Encryption.header` stays the old one until the commit, because that is what opens the items readers get. Added after the first draft of v1."
          }
        }
      },
      "Encryption": {
        "type": "object",
        "required": [
          "state"
        ],
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "plaintext",
              "sealed",
              "rewriting"
            ]
          },
          "keyId": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/KeyId"
              },
              {
                "type": "null"
              }
            ],
            "description": "While a rewrite is open: still the key readers need, the one before it."
          },
          "header": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Opaque"
              },
              {
                "type": "null"
              }
            ]
          },
          "rewrite": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/RewriteSession"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "Workspace": {
        "type": "object",
        "required": [
          "name",
          "quotaBytes",
          "usedBytes",
          "itemCount",
          "encryption"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "quotaBytes": {
            "$ref": "#/components/schemas/Bytes"
          },
          "usedBytes": {
            "$ref": "#/components/schemas/Bytes"
          },
          "itemCount": {
            "type": "integer",
            "minimum": 0
          },
          "encryption": {
            "$ref": "#/components/schemas/Encryption"
          }
        }
      },
      "Viewer": {
        "type": "object",
        "required": [
          "key",
          "server"
        ],
        "properties": {
          "key": {
            "type": "object",
            "required": [
              "id",
              "role"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "label": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "role": {
                "type": "string",
                "enum": [
                  "readWrite",
                  "readOnly"
                ]
              },
              "expiresAt": {
                "oneOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          },
          "server": {
            "type": "object",
            "required": [
              "version",
              "apiVersion"
            ],
            "properties": {
              "version": {
                "type": "string"
              },
              "apiVersion": {
                "type": "integer"
              },
              "maxItemBytes": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Bytes"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "null: no limit. The client checks it before it uploads, because its other backends have none."
              },
              "sourceUrl": {
                "type": "string",
                "format": "uri",
                "description": "Where the source of this server is. The server is AGPL-3.0-or-later: whoever runs a modified one for others must offer them its source, and points this at it. Added after the first draft of v1; a client must not require it."
              }
            }
          }
        }
      },
...
      "CleanStagingRequest": {
        "type": "object",
        "required": [
          "olderThanSecs"
        ],
        "properties": {
          "olderThanSecs": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "CleanStagingResult": {
        "type": "object",
        "required": [
          "removed"
        ],
        "properties": {
          "removed": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "CONTENT_MISMATCH",
          "FORBIDDEN_ROLE",
          "INVALID_ID",
          "INVALID_REQUEST",
          "ITEM_TOO_LARGE",
          "KEY_EXPIRED",
          "KEY_ID_MISMATCH",
          "KEY_REVOKED",
          "LEASE_HELD",
          "NOT_FOUND",
          "QUOTA_EXCEEDED",
          "RATE_LIMITED",
          "REWRITE_ENDED",
          "REWRITE_INCOMPLETE",
          "REWRITE_IN_PROGRESS",
          "SERVICE_UNAVAILABLE",
          "UNAUTHENTICATED"
        ]
      },
      "Problem": {
        "type": "object",
        "required": [
          "code",
          "status",
          "title"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "leaseExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "RFC 9457 problem details, with a stable `code`."
      }
    }
  }
}